requirements.txt
certifi==2023.7.22; python_version >= '3.6'
charset-normalizer==3.2.0; python_version >= '3.7'
idna==3.4; python_version >= '3.5'
jinja2==3.1.2
markupsafe==2.1.3; python_version >= '3.7'
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.4; python_version >= '3.7'